+++ generators/app/index.js
@@ -129,9 +129,6 @@
 
 			utils.installNpmPackages(this);
 			utils.installBowerPackages(this);
-			this.spawnCommand('npm', [
-				'install'
-			]);
 		}
 
 		end() {